Move widget object files from projects/components to projects/widget-object
