+++ projects/widget-object/src/lib/angular/angular-widget-finder.ts
