+++ projects/widget-object/src/lib/angular/angular-widget-object-element.ts
