+++ projects/widget-object/src/lib/angular/angular-widget-object.spec.ts
