+++ projects/widget-object/src/lib/index.ts
