+++ projects/widget-object/src/lib/selector-util.ts
