--- projects/widget-object/src/lib/widget-object.spec.ts
@@ -1,11 +1,0 @@
-/*!
- * Copyright 2021 VMware, Inc.
- * SPDX-License-Identifier: BSD-2-Clause
- */
-import { widgetObject } from './widget-object';
-
-describe('widgetObject', () => {
-    it('should work', () => {
-        expect(widgetObject()).toEqual('widget-object');
-    });
-});
